system: core: adb: Change StartLimitBurst to 10 for adbd
On some platforms, when we perform a usb composition switch in a
loop, then sometimes we end up requesting SYSTEMD to restart the
adbd more than the default startlimtburst(5) in the the default
startlimitintervalsec(10s) time frame.
In such case, SYSTEMD will not resart the adbd, mark it as failed
due to start-limit-hit and the usb composition switch fails.
Fix this by changing the startlimitburst to 10 for adbd.service
only as we anyway sleep for 1s after killing adbd in the
individual usb compositions.

Change-Id: I73d2e9acf750921818d5866908c9e20462c4b5ec
Signed-off-by: Harsh Agarwal <harshq@codeaurora.org>