system: core: Delay enum on sdxnightjar until uC is loaded
When sdx12 is connected to ubuntu pc by setting default
as a tethering composition and device is rebooted, ubuntu host
sends a sequence of set alt 1 0 1 which calls ipa_xdci_connect,
disconect and again connect. These happen in a queue context.
ipa_xdci_disconnect on sdxnightjar is dependant on uC loading,
hence currently ipa_xdci_disconnect call waits for uC to be
loaded which takes around 8-9 seconds and timeout occurs
at host end. This results in usb enumeration failure.

This can happen not only with ubuntu host but with any host
if sequence of set alt 1 0 1 is sent. The reason this change
is made for all compositions is that if usb is enumerated in
any non tethering composition and if composition switch occurs
to tethering composition without uC loaded and if host driver
is programmed to send a sequence of set alt 1 0 1, then this
results in enumeration failure. To solve this delay usb enumeration
on sdxnightjar until uC is loaded.

Change-Id: Id7f011c9d2119cb814d364fae3ab827e2af458f4
Signed-off-by: Rohith Kollalsi <quic_rkollals@quicinc.com>